==> R/00_main.qmd <==
---
title: "00_main"
format: html
editor: visual
---

```{r}
library(quarto)
```

```{r}
quarto::quarto_render("./01_load.qmd")
quarto::quarto_render("./02_clean.qmd")
quarto::quarto_render("./03_annotation.qmd")
quarto::quarto_render("./04_log2_transformation.qmd")
quarto::quarto_render("./05_describe.qmd")
quarto::quarto_render("./06_visualise_expr_significance.qmd")  
quarto::quarto_render("./07_PCA.qmd") 
quarto::quarto_render("./08_GO_analysis.qmd")
quarto::quarto_render("./09_Chromosomal_distribution.qmd") 
quarto::quarto_render("./10_Read_count_analysis.qmd") 
quarto::quarto_render("./presentation.qmd")

```
