==> R/04_log2_transformation.qmd <==
---
title: "04_augment"
format: html
editor: visual
---

## Load libraries

```{r}
# Output: FALSE
library(tidyverse)
library(broom)
```

## Load data

```{r}
# Output: FALSE
# Loading the annotated gene dataset
data = read_tsv(gzfile("../data/annotated_meta_gene_table.tsv.gz"))
```

## log2 transformation of expression

To reduce the range of expression variance while still observing gene expression differences, we applied a log2 transformation to the data, following the method used in the original paper.

```{r}
aug_data <- data |>
   mutate(log2_Lesion = 
           if_else(Lesion == 0, 
                   0, 
                   log2(Lesion)
                   )
         ) |> 
   mutate(log2_Perilesion = 
           if_else(Perilesion == 0, 
                   0, 
                   log2(Perilesion)
                   )
         ) |> 
   mutate(log2_expr_change = 
           if_else(expr_diff == 0, 
                   0, 
                   log2(expr_diff)
                   )
         ) 

```

## Write tsv file

```{r}
# We save the augmented data frame
write_tsv(aug_data, 
          file = gzfile("../data/augmented_meta_gene_table.tsv.gz"))
```
