==> R/05_describe.qmd <==
---
title: "05_describe"
format: html
project:
    output-dir: '../Docs'
editor: visual
bibliography: references.bib
---

```{r}
library(tidyverse)
```

# Data description

```{r}
# load in augmented data for description
data = read_tsv(gzfile('../data/augmented_meta_gene_table.tsv.gz'))
```

# Samples

Hidradenitis suppurativa (HS) is a severe, debilitating chronic inflammatory disease affecting hair follicles with a predilection for intertriginous areas.[@freudenberg2023]. We looked at this study of 20 HS patients at the University of North Carolina Department of Dermatology clinics from October 2017 through February 2018. Whole skin punch biopsies were collected from a lesional site and a matching non-lesional site located 5 cm from the site of inflammation for all the patients[@freudenberg2023]. We want to look at the genes that are significantly differentially expressed between the lesional and perilesional sites; the idea is that the genes that are significantly over/under-expressed in the lesional sites might be involved in molecular pathways that affect HS.

```{r}
meta_cols <- c('Sample',
               'Gender',
               'Race',
               'Smoker',
               'Age',
               'BMI')

```

## Select significant genes

```{r}

# Data description like number of genes, number of patients etc.
data |> 
  dplyr::select(gene) |> 
  unique() |> 
  nrow()
```

There are 453 genes we are looking at across the lesional and perilesional tissue samples.

## Gender

```{r}
wide_data <- 
  data |>
  dplyr::select(all_of(meta_cols),
         gene,
         Lesion,
         Perilesion) |> 
  pivot_longer(c(Lesion, Perilesion),
               names_to = 'Tissue_type',
               values_to = 'expression') |> 
  pivot_wider(names_from = gene, 
              values_from = expression)

wide_data |> 
  filter(Tissue_type == 'Perilesion') |> 
  group_by(Gender) |> 
  count() |> 
  ungroup()
```

Out of the 20 patients selected for this study, 17 were females and 3 were males.\

## Race

```{r}
wide_data |>
  filter(Tissue_type == "Lesion") |> 
  group_by(Race) |> 
  summarise(n = n())
```

##  Smoking habits

```{r}
# Grouping by gender and smoking history
wide_data |>
  dplyr::filter(Tissue_type == "Lesion") |> 
  group_by(Smoker, Gender) |>
  summarize(n = n()) |> 
  ungroup()
```

6 patients were ex-smokers, 9 were non-smokers and 5 were smokers.

5 female patients were ex-smokers, 9 were non-smokers and 3 were smokers.

1 male patient was an ex-smoker, and 2 were smokers. So no male patient was a non-smoker.\

```{r}
attribute_dist_table <- wide_data |> 
  dplyr::select(Sample, Age, Gender, BMI, Smoker, Race)

attribute_dist_table
```

```{r}
# Creating age group

attribute_dist_table <- 
  attribute_dist_table |>  
  mutate(Age_group = case_when(
    Age >= 15 & Age <= 20 ~ '[15:20]',
    Age > 20 & Age <= 30 ~ '(20:30]',
    Age > 30 & Age <= 40 ~ '(30:40]',
    Age > 40 & Age <= 50 ~ '(40:50]',
    Age > 50 & Age <= 60 ~ '(50:60]'
  )) |> 
  mutate(Age_group = factor(
    Age_group, 
    levels = c('[15:20]', 
               '(20:30]', 
               '(30:40]', 
               '(40:50]', 
               '(50:60]')
    ))
```

```{r}
# Write out categories
attribute_dist_table <- 
  attribute_dist_table |> 
  mutate(Gender = if_else(Gender == 'F',
                          'Female',
                          'Male')
         ) |>
  mutate(Smoker = case_when(Smoker == 'E'~'Ex-smoker',
                            Smoker == 'Y'~'Yes',
                            Smoker == 'N'~'No')
         )

attribute_dist_table
```

##  Age distribution by Gender

```{r}
attribute_dist_table |>
  ggplot(aes(x = Age_group,
             fill = Gender
             )
         ) +
  geom_bar(color = "black", 
           alpha = 0.5) +
  geom_hline(yintercept = 0,
             color = 'black',
             size = 0.5) +
  facet_wrap(~Gender) +
  labs(title = 'Age Distribution by Gender',
       x = 'Age group',
       y = 'n') +
  theme_minimal() +
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        strip.text = element_text(size = 11), # Adjust facet title size
        ) +
  guides(fill = "none") # Remove legend

```

##  Age Distribution by Smoking history and Gender

```{r}
attribute_dist_table |>
  ggplot(aes(x = Age_group,
             fill = Gender
             )
         ) +
  geom_bar(color = "black", 
           alpha = 0.5) +
  geom_hline(yintercept = 0,
             color = 'black',
             size = 0.5) +
  facet_wrap(~Smoker) +
  labs(title = 'Age Distribution by Smoking history and Gender',
       x = 'Age group',
       y = 'n') +
  theme_minimal() +
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        strip.text = element_text(size = 11),
        axis.text.x = element_text(angle = 45,
                                   hjust = 1,
                                   vjust = 1.5)
        )
```

```{r}
# Add BMI group
attribute_dist_table <- attribute_dist_table |>  
  mutate(BMI_group = case_when(BMI <= 25 ~ 'Below 25',
                               (BMI > 25 & BMI <= 35) ~ 'Between 25 and 35',
                               (BMI > 35 & BMI <= 45) ~ 'Between 35 and 45',
                               BMI > 45 ~ 'Above 45')
         ) |> 
  mutate(BMI_group = factor(BMI_group, 
                            levels = c('Below 25', 
                                       'Between 25 and 35', 
                                       'Between 35 and 45', 
                                       'Above 45')
                            ))

```

##  BMI Distribution by Gender

```{r}
attribute_dist_table |>
  ggplot(aes(x = BMI_group,
             fill = Gender
             )
         ) +
  geom_bar(color = "black", 
           alpha = 0.5) +
  geom_hline(yintercept = 0,
             color = 'black',
             size = 0.5) +
  facet_wrap(~Gender) +
  labs(title = 'BMI Distribution by Gender',
       x = 'BMI group',
       y = 'n') +
  theme_minimal() +
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        strip.text = element_text(size = 11), # Adjust facet title size
        axis.text.x = element_text(angle = 40,
                                   hjust = 1,
                                   vjust = 1.1)
        ) +
  guides(fill = "none") # Remove legend
```

## BMI Distribution by Smoking history and Gender

```{r}
attribute_dist_table |>
  ggplot(aes(x = BMI_group,
             fill = Gender
             )
         ) +
  geom_bar(color = "black", 
           alpha = 0.5) +
  geom_hline(yintercept = 0,
             color = 'black',
             size = 0.5) +
  facet_wrap(~Smoker) +
  labs(title = 'BMI Distribution by Smoking history and Gender',
       x = 'BMI group',
       y = 'n') +
  theme_minimal() +
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        strip.text = element_text(size = 11),
        axis.text.x = element_text(angle = 40,
                                   hjust = 1,
                                   vjust = 1.1)
        )
```

##  BMI Distribution by Race

```{r}
attribute_dist_table |>
  ggplot(aes(x = BMI_group,
             fill = Race
             )
         ) +
  scale_fill_manual(values = c("B" = "cyan", "O" = "salmon", "W" = "violet", "W_A" = "yellow"),
                    labels = c("Black", "Oriental", "White", "White, Asian")) +
  geom_bar(color = "black", 
           alpha = 0.5) +
  geom_hline(yintercept = 0,
             color = 'black',
             size = 0.5) +
  facet_wrap(~Race) +
  labs(title = 'BMI Distribution by Race',
       x = 'BMI group',
       y = 'n') +
  theme_minimal() +
  theme(panel.grid = element_blank(),
        plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        strip.text = element_text(size = 11),
        axis.text.x = element_text(angle = 40,
                                   hjust = 1,
                                   vjust = 1.1)
        )
  #guides(fill = 'none')
```
