==> R/08_GO_analysis.qmd <==
---
title: "GO_analysis"
format: html
editor: visual
---

## Gene Ontology Analysis

We performed Gene Ontology analysis aimed at identifying altered biological pathways, we used **`enrichGO`** from the **`clusterProfiler`** package from Bioconductor.

## Load libraries

```{r}
# ouput: FALSE
library(tidyverse)
library(clusterProfiler)
library(AnnotationDbi)
library(org.Hs.eg.db)
```

## Load data

```{r}
# output: FALSE
expression_data <- read_tsv(gzfile('../data/augmented_meta_gene_table.tsv.gz'))
```

## Gene Ontology analysis for genes with upregulated expression

We initially look at which Biological Pathway was mostly upregulated based on the upregulated gene. We also had the option to look at the Cellular Components (CC) and Molecular Functions (MF), but we deemed those less relavant for our pupose.

```{r}
upregulated <- 
  expression_data |> 
  group_by(gene) |> 
  filter(any(expr_diff >= 3)) |> 
  pull(gene) 


upreg_GO_res <- 
  enrichGO(gene = upregulated,
                   OrgDb = 'org.Hs.eg.db',
                   keyType = 'ENSEMBL',
                   ont = 'BP') |>  # BP for biological processes
          as_data_frame()
          
# Shorten the string, so the description does not overlap in the visualisation
upreg_GO_res <- 
  upreg_GO_res |> 
   mutate(short_description = ifelse(str_length(Description) > 20,
                                   str_c(str_sub(Description, 1, 20), "...", ""),
                                   Description))  
```

```{r fig.height=12, fig.width = 8}

ureg_overview <- 
  upreg_GO_res |> 
  # Order x based on ascending p.adj
  ggplot(aes(x = fct_reorder(short_description, 
                             p.adjust), 
             y = Count, fill = p.adjust)) + 
  # stat = 'identity' tells y is the height of bars
  # position = 'dodge' basicallty says that the x-axis is catagorical
  geom_bar(stat = "identity", 
           position = "dodge") +
  scale_fill_gradient(low = "blue", 
                      high = "red") +  
  labs(
    title = "Differential Expression Profile: Top Gene Ontology Terms",
    x = "Gene Ontology Term Description", 
    y = "Count",
    fill = "Adjusted p-value"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5,
                              face = 'bold'),
    axis.text.x = element_text(size = 12),
    axis.text.y = element_text(size = 12),
    panel.border = element_rect(colour = "black", 
                                fill = 'transparent',
                                size = 0.5) 
        ) +
  # Flip coordinates
  coord_flip()  

ureg_overview 
ggsave("../pictures/GO_upreg_overview.png", ureg_overview)
```

## Only for top 20 pathways

Due to a large number of upregulated pathways identified, we decided to narrow our focus to the top 20 pathways. This prioritization was based on the adjusted p-values (p.adj), which indicate the confidence level in the association of the identified gene groups with each specific pathway.

```{r fig.height = 7, fig.width = 6}


upreg_GO_res <- 
  upreg_GO_res |> 
   mutate(short_description = ifelse(str_length(Description) > 30,
                                   str_c(str_sub(Description, 1, 30), "..."),
                                   Description)) |>    
   mutate(short_description = str_wrap(short_description, 
                                      width = 20)) 


upreg_top20 <- upreg_GO_res |> 
  arrange(desc(p.adjust)) |>
  dplyr::slice(1:20) |> 
  ggplot(aes(x = fct_reorder(short_description, p.adjust), 
             y = Count, 
             fill = p.adjust)) +  
  # stat = 'identity' tells y is the height of bars
  # position = 'dodge' basicallty says that the x-axis is catagorical
  geom_bar(stat = "identity", 
           position = "dodge") +
  scale_fill_gradient(low = "blue", 
                      high = "red") +  
  labs(
    title = "Differential Expression Profile: Top 20 Gene Ontology Terms",
    x = "Gene Ontology Term Description", 
    y = "Count",
    fill = "Adjusted p-value"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5,
                              face = 'bold'),
    axis.text.x = element_text(size = 11),
    axis.text.y = element_text(size = 11),
    axis.title.y = element_blank(),
    panel.border = element_rect(colour = "black", fill = 'transparent',size = 0.5) 
        ) +
  # Flip coordinates
  coord_flip()  

upreg_top20
ggsave("../pictures/GO_upreg_top20.png", upreg_top20, bg = "white")
```

## Gene Ontology analysis for downregulated gene

Same as above, just for downregulated genes

```{r}
downregulated <- 
  expression_data |> 
  group_by(gene) |> 
  filter(any(expr_diff <= 1/3)) |> 
  pull(gene) 



downreg_GO_res <- enrichGO(gene = downregulated,
                   OrgDb = 'org.Hs.eg.db',
                   keyType = 'ENSEMBL',
                   # BP for biological processes
                   ont = 'BP') |>  
          as.data.frame()


downreg_GO_res <- 
  downreg_GO_res |> 
   mutate(short_description = ifelse(str_length(Description) > 30,
                                   str_c(str_sub(Description, 1, 30), "...", ""),
                                   Description)) |> 
  mutate(short_description = str_wrap(short_description, 20))  
```

```{r fig.height = 7, fig.width = 6}


downreg_overview <- downreg_GO_res |> 
  ggplot(aes(x = fct_reorder(short_description, 
                             p.adjust), 
             y = Count, 
             fill = p.adjust)) +  
  # stat = 'identity' tells y is the height of bars
  # position = 'dodge' basicallty says that the x-axis is catagorical
  geom_bar(stat = "identity", 
           position = "dodge") +
  scale_fill_gradient(low = "blue", 
                      high = "red") +  
  labs(
    title = "Differential Expression Profile: Bottom Gene Ontology Terms",
    x = "Gene Ontology Term Description", 
    y = "Count",
    fill = "Adjusted p-value"
  ) +
  theme_minimal() +
  theme(
    plot.title = element_text(hjust = 0.5,
                              face = 'bold'),
    axis.text.x = element_text(size = 11),
    axis.text.y = element_text(size = 11),
    axis.title.y = element_blank(),
    panel.border = element_rect(colour = "black", fill = 'transparent',size = 0.5) 
        ) +
  coord_flip()  # Flip coordinates to make the bars horizontal

downreg_overview
ggsave("../pictures/GO_downreg_overview.png", downreg_overview, bg = "white")
```
