==> R/09_Chromosomal_distribution.qmd <==
---
title: "Chromosomal_distribution"
format: html
editor: visual
---

```{r}
library(tidyverse)
```

```{r}
#Load in our significant genes that are either up or downregulated
expression_data <- read_tsv(gzfile('../data/augmented_meta_gene_table.tsv.gz'))
```

```{r}
# Rearrange data
arranged_data <- 
  expression_data |> 
  dplyr::select(-c(Lesion, Perilesion)) |> 
  pivot_longer(c(log2_Lesion, log2_Perilesion),
               names_to = 'Tissue_type',
               values_to = 'log2expr') |> 
  mutate(Tissue_type = if_else(Tissue_type == 'log2_Lesion',
                               'Lesion',
                               'Perilesion')) |> 
  group_by(gene) |> 
  mutate(mean_log2expr = mean(log2expr)) |> 
  mutate(CHR = factor(CHR, levels = c(paste(1:22), "X", "Y")))
  
```

## Count of up- and downregulated genes

```{r}
upreg_count <- 
  arranged_data |>
  group_by(gene) |> 
  # Only look at the upregulated genes
  filter(any(expr_diff >= 3)) |> 
  dplyr::select(gene) |> 
  unique() |> 
  nrow()

downreg_count <- 
  arranged_data |>
  group_by(gene) |> 
  # Only look at the downregulated genes
  filter(any(expr_diff <= 1/3)) |> 
  dplyr::select(gene) |> 
  unique() |> 
  nrow()

gene_counts <- 
  data.frame('Number_of_upregulated_genes' = upreg_count,
           'Number_of_downregulated_genes' = downreg_count)

gene_counts
```

## Chromosome distribution of upregulated genes

```{r}
chromosomal_dist_upreg <- arranged_data |>
  group_by(gene) |> 
  # Only look at the upregulated genes
  filter(any(expr_diff >= 3)) |> 
  ungroup() |> 
  # emphasize sex chromosoms
  mutate(sex_chromosom = ifelse((CHR == "X" | CHR == "Y"), 1, 0)) |>
  # fct_rev is used to filp the CHR catagory
  ggplot(aes(fct_rev(CHR), fill = sex_chromosom)) +
  geom_bar() +
  coord_flip() +
  theme_minimal() +
  labs(title = 'Chromosomal distribution of upregulated genes',
       x = 'Chromosome',
       y = 'Number of upregulated genes') +
  theme(plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        legend.position = "none",
        panel.border = element_rect(colour = "black", 
                                    fill = 'transparent',
                                    size = 0.5),
        axis.text = element_text(size = 10)
        )

chromosomal_dist_upreg

ggsave("chromosomal_dist_upreg.png", 
       chromosomal_dist_upreg, 
       path = "../pictures", 
       bg = "white")
```

## Chromosomal distribution of downregulated genes

```{r}
chromosomal_dist_downreg <- arranged_data |>
  group_by(gene) |> 
  # Only look at the downregulated genes
  filter(any(expr_diff <= 1/3)) |> 
  ungroup() |>  
  # emphasize sex chromosoms
  mutate(sex_chromosom = ifelse((CHR == "X" | CHR == "Y"), 1, 0)) |>
  # fct_rev is used to filp the CHR catagory
  ggplot(aes(fct_rev(CHR), fill = sex_chromosom)) +
  geom_bar() +
  coord_flip() +
  theme_minimal() +
  labs(title = 'Chromosomal distribution of downregulated genes',
       x = 'Chromosome',
       y = 'Number of downregulated genes') +
  theme(plot.title = element_text(hjust = 0.5,
                                  vjust = 2,
                                  face = 'bold'),
        legend.position = "none",
        panel.border = element_rect(colour = "black", 
                                    fill = 'transparent',
                                    size = 0.5),
        axis.text = element_text(size = 10)
        )

chromosomal_dist_downreg

ggsave("chromosomal_dist_downreg.png", 
       chromosomal_dist_downreg, 
       path = "../pictures",
       bg = "white")
```
